requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncpg==0.29.0
attrs==23.2.0
Authlib==1.3.1
bcrypt==4.1.3
black==24.4.2
boto3==1.34.139
botocore==1.34.139
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastapi-utils==0.7.0
frozenlist==1.4.1
greenlet==3.0.3
grpcio==1.65.1
grpcio-tools==1.65.1
h11==0.14.0
hiredis==2.4.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
loguru==0.7.2
magic-filter==1.0.12
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mmh3==4.1.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==2.0.0
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
platformdirs==4.2.2
protobuf==5.27.2
psutil==5.9.8
pycparser==2.22
pydantic>=2.4.1
pydantic-settings==2.3.4
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-ulid==2.7.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
requests==2.32.3
rich==13.7.1
s3transfer==0.10.2
setuptools==71.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqladmin==0.18.0
SQLAlchemy==2.0.31
starlette==0.37.2
tqdm==4.66.4
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
WTForms==3.1.2
yarl==1.9.4
clickhouse_connect==0.7.18
humanize~=4.10.0
aiosmtplib~=3.0.2
pytest==8.3.3
jsonschema==4.23.0
isort==5.13.2
allure-pytest==2.13.5
curlify2==2.0.0
deepdiff==8.0.1
openai~=1.47.0
allure-python-commons~=2.13.5